refactor: move OpenFeatureAPIFactory to sdk package and remove createIsolated()
